Extracts the SASS generator into its own class for better testability.
Fixes: SIRI-622
